requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.0
attrs==20.3.0
auditlog3==1.0.1
black==20.8b1
certifi==2020.12.5
click==7.1.2
coverage==5.3
Django==3.0.10
django-auditlog==0.4.7
django-cors-headers==3.6.0
django-environ==0.4.5
django-extensions==3.0.9
django-extra-fields==3.0.2
django-filter==2.4.0
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-mptt==0.11.0
django-recurrence==1.10.3
djangorestframework==3.11.1
elastic-apm==5.10.0
flake8==3.8.4
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.6.4
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
promise==2.3
psycopg2==2.8.6
py==1.9.0
pycodestyle==2.6.0
pydot==1.4.1
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.6.0      # resolved version due to django-auditlog, probably not safe to update
pytz==2020.1
regex==2020.10.28
Rx==1.6.1
sentry-sdk==0.19.5
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.26.2
uWSGI==2.0.19.1
whitenoise==5.2.0
zipp==3.4.0
django-modeltranslation==0.16.1
